pyproject.toml
[tool.poetry]
name = "rlt-mongo"
version = "0.1.0"
description = ""
authors = ["kylych-dev <mirbekov1kylych@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymongo = "^4.7.2"
motor = "^3.4.0"
aiogram = "^3.6.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
